Added observer's notification, changed IterationData.Builder

+++ src/main/java/pl/edu/agh/bo/knabees/communication/IterationData.java
@@ -15,7 +15,7 @@ public final class IterationData {
 
 		private double totalTakenValue;
 
-		public Builder(int iterationNumber, int totalTakenValue) {
+		public Builder(int iterationNumber, double totalTakenValue) {
 			this.iterationNum = iterationNumber;
 			this.totalTakenValue = totalTakenValue;
 		}
